+++ webpage-to-markdown-local/src/webpage-to-offline-markdown.tsx
@@ -1,5 +1,5 @@
 import { WebpageToMarkdownCommand } from "./command";
 
 export default function Command() {
-  return <WebpageToMarkdownCommand />;
+  return <WebpageToMarkdownCommand offline />;
 }